fix(HotelRatingService): 评论配额应当考虑订单的amount

+++ backend/base/src/infrastructure/repository/transaction.rs
@@ -402,7 +402,7 @@ impl OrderDataConverter {
                 .map(TransactionId::try_from)
                 .transpose()?,
         );
-        let personal_info_id = PersonalInfoId::from_db_value(dish_order_do.person_info_id as i32)?;
+        let personal_info_id = PersonalInfoId::from_db_value(dish_order_do.person_info_id)?;
 
         let base = BaseOrder::new(
             Some(order_id),
